Add mask docs, finalize coverage/tests
