--- app/ids-mask/function.html
@@ -1,15 +1,0 @@
-{{> ../layouts/head.html }}
-
-<ids-layout-grid cols="3" gap="md">
-  <ids-layout-grid-cell>
-    <ids-text font-size="12" type="h1">Ids Input (Masked) - Custom Mask Function</ids-text>
-  </ids-layout-grid-cell>
-</ids-layout-grid>
-
-<ids-layout-grid cols="3" gap="md">
-  <ids-layout-grid-cell>
-    <ids-input id="mask-custom" label="Ed Only" placeholder="Ed"></ids-input>
-  </ids-layout-grid-cell>
-</ids-layout-grid>
-
-{{> ../layouts/footer.html }}
