+++ src/ids-input/ids-input.js
@@ -129,13 +129,6 @@ class IdsInput extends mix(IdsElement).with(...appliedMixins) {
    */
   connectedCallback() {
     super.connectedCallback?.();
-
-    /*
-    if (this.value === null) {
-      this.value = '';
-    }
-    */
-
     this.handleEvents();
     this.handleAutoselect();
     this.handleClearable();
